requirements.txt
gradio==5.45.0
python-dotenv
openai
requests
base64
json